// src/main/resources/static/project-control-ui/src/app/api/model/entityModelAlias.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { GitRepo } from './gitRepo';
import { Link } from './link';


export interface EntityModelAlias { 
    repo?: GitRepo;
    id?: number;
    name?: string;
    user?: User;
    _links?: { [key: string]: Link; };
}